double parameters

+++ src/main/java/com/qa/calculator/Add.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Add implements Editable {
 
-	public double editNumbers(int num1, int num2) {
+	public double editNumbers(double num1, double num2) {
 		return num1 + num2;
 	}
 }
